+++ changelog.md
@@ -1,8 +1,11 @@
 ==== 1.3.3 ====
 
+  * ADDED: Spoofing Card for OpenComputers
+    - Works the same as a Network Card, but allows you to specify the source address
   * ADDED: OpenPeripheral integration:
     - Now OpenPeripheral peripherals aren't ignored anymore if Computronics is present
   * FIXED: Lamps ignoring bundled redstone if placed adjacent to each other
+  * MISC: Added a few more things to the config file
 
 ==== 1.3.2 ====